sideBar components separated, names changed
