+++ website/components/sideBar/sideBar.module.css
@@ -13,5 +13,5 @@
 }
 
 .topElement:hover {
-    color: #053aad;
+    color: #0578ad;
 }
